refactor: messageSendBox 하나로 통일
